Fixes #15148: Add search/pagination to eventlogs UI
