+++ webapp/sources/rudder/rudder-web/src/main/scala/bootstrap/liftweb/Boot.scala
@@ -284,6 +284,7 @@ class Boot extends Loggable {
     LiftRules.statelessDispatch.append(RudderConfig.restQuicksearch)
     LiftRules.statelessDispatch.append(RudderConfig.restCompletion)
     LiftRules.statelessDispatch.append(RudderConfig.sharedFileApi)
+    LiftRules.statelessDispatch.append(RudderConfig.eventLogApi)
 
     // REST API (all public/internal API)
     // we need to add "info" API here to have all used API (even plugins)
